refactor(canvas): colocate chat input components
